% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewline.R
\name{add_newline}
\alias{add_newline}
\title{Add newlines to a vector of Latex file lines, used in cases of section/subsection, 
or double dollar symbol for math equation}
\usage{
add_newline(x, grep_ptn, paired = "none", pos = "after")
}
\arguments{
\item{x}{a vector of Latex file lines}

\item{grep_ptn}{match pattern for function "grep"}

\item{paired}{character, specify paired status c("none","start","end"),
the default "paired = 'none'"}

\item{pos}{character, where to add newline c("before","after"),
the default "pos = 'after'"}
}
\value{
vector
}
\description{
Add newlines to a vector of Latex file lines, used in cases of section/subsection, 
or double dollar symbol for math equation
}
\details{
use this function to add newlines in case section/subsection, 
or double dollar symbol for math equation
}
\examples{
\dontrun{
 ## add newline before "$$" by using custom function
tex <- add_newline(          
 x = tex,  grep_ptn = "\\\\$\\\\$",
 paired = "start",
 pos = "before")
## add newline after "$$" by using custom function
tex <- add_newline(          
 x = tex,  grep_ptn = "\\\\$\\\\$",
 paired = "end",
 pos = "after")
}

}
