% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTexTag.r
\name{convertTexTag}
\alias{convertTexTag}
\title{convertTexTag}
\usage{
convertTexTag(x, tagFrom, mdTo)
}
\arguments{
\item{x}{A vector of strings containing lines from the Latex file}

\item{tagFrom}{The Latex tag to convert from}

\item{mdTo}{The string to surround argument of Latex command with.}
}
\description{
Convert a tex tag, like \\texttt, to something else for the RMD file
}
