% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBib.r
\name{parseBib}
\alias{parseBib}
\title{parseBib}
\usage{
parseBib(x)
}
\arguments{
\item{x}{A latex bibitem entry from a Latex .bbl file}
}
\value{
A string containing the bibitem
}
\description{
Parse a bibitem and return the bibliography item
}
\examples{
bibItem <- "\\\\bibitem[{Bai et~al.(2011)Bai, L{\\\\\"{u}},
Wang, Zhou and Ding}]{Bai2011a} Bai, S., L{\\\\\"{u}}, G.,
Wang, J., Zhou, P. and Ding, L. (2011)
{GIS-based rare   events logistic regression for
landslide-susceptibility mapping of   Lianyungang, China}.
\\\\newblock \\\\textit{Environmental Earth Sciences},
\\\\textbf{62}, 139--149. "

parseBib(bibItem)

}
\author{
Trent McDonald
}
