% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCrossRefs.r
\name{processCrossRefs}
\alias{processCrossRefs}
\title{processCrossRefs}
\usage{
processCrossRefs(tex, fileRoot)
}
\arguments{
\item{tex}{A character vector representing the lines of a Latex file.}

\item{fileRoot}{The root name of the Latex file. This function
looks for an reads the file fileRoot.aux.}
}
\value{
A character vector representing the lines of a Latex file,
but with all the cross references fixed up.
}
\description{
Process the cross references in a Latex file, substituting
the correct number in for all the \\ref{} tags.
}
