% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDisplayEqns.r
\name{processDisplayEqns}
\alias{processDisplayEqns}
\title{Process the display equations in the tex file}
\usage{
processDisplayEqns(tex)
}
\arguments{
\item{tex}{An array of latex lines}
}
\value{
A \code{tex} array, just like the input, with
display equations formated so they work in markdown.
}
\description{
Processes all the display equations, converting them
into a markdown format.
}
