% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFigures.r
\name{processFigures}
\alias{processFigures}
\title{processFigures}
\usage{
processFigures(x)
}
\arguments{
\item{x}{A vector of lines from the Latex file}
}
\value{
A vector of lines with latex figures replace by something
that will render a Rmd Figure
}
\description{
Process the figures in a Latex file vector and
put into Rmd format
}
