% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processListTypes.r
\name{processListTypes}
\alias{processListTypes}
\title{processOneListType - Process one type of list environment}
\usage{
processListTypes(tex, listTypes)
}
\arguments{
\item{tex}{a vector of Latex file lines}

\item{listType}{A vector containing the types lists to process.
For example, c("itemize", "enumerate").  Note that all list types
must be process-able in function \code{processOneList}.}
}
\value{
a processed vector of Latex file lines with lists
formated in R Markdown.
}
\description{
Process one \code{itemize} or \code{enumerate} list
environment.
}
