% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processNewLines.r
\name{processNewLines}
\alias{processNewLines}
\title{processNewLines - Process manual new lines}
\usage{
processNewLines(tex)
}
\arguments{
\item{tex}{a vector of Latex file lines}
}
\value{
a processed vector of Latex file lines with lists
formated in R Markdown.
}
\description{
Process (remove) double blanks at the end of
a line which causes R Markdown to start a new line, and
replace double backslashes or newline commands with double blanks to force
a new line.
}
