% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processOneFigure.r
\name{processOneFigure}
\alias{processOneFigure}
\title{processOneFigure}
\usage{
processOneFigure(x, begin, end, figNum)
}
\arguments{
\item{x}{A vector of latex lines}

\item{begin}{The beginning line number in the overall Latex
file of the figure to process}

\item{end}{The ending line number in the overall Latex file
of the figure to process}
}
\description{
Process one latex Figure into Rmd format
}
\details{
The \\begin{Figure} and \\end{Figure} command must be
on lines by themselves.
}
