% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processOneList.r
\name{processOneList}
\alias{processOneList}
\title{processOneList - Process one list environment}
\usage{
processOneList(tex, start, end, listType, listLevel)
}
\arguments{
\item{tex}{a vector of Latex file lines}

\item{start}{line number where list starts (begin statement)}

\item{end}{line number where list ends (end statement)}

\item{listType}{The type of list to process. Currently, only "itemize"
or "enumerate".}

\item{listLevel}{the level of the list.  0 is top level, 1 is for
lists embedded in a 0 level list (i.e., 2nd level), etc.}
}
\value{
a processed vector of Latex file lines with lists
formated in R Markdown.
}
\description{
Process one \code{itemize} or \code{enumerate} list
environment.
}
