% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processOneTable.r
\name{processOneTable}
\alias{processOneTable}
\title{processOneTable}
\usage{
processOneTable(x, begin, end, tabNum)
}
\arguments{
\item{x}{A vector of latex lines}

\item{begin}{The beginning line number of the latex table to process}

\item{end}{The ending line number of the latex table to process}
}
\description{
Process one latex table into Rmd format
}
\details{
The \\begin{table} and \\end{table} command must be on lines by themselves.

The header of a latex table is signified by the first \\hline, or the
last row with a \\multicolumn command.  Only one-line headers are
allowed in Markdown.
}
