% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTables.r
\name{processTables}
\alias{processTables}
\title{processTables}
\usage{
processTables(x)
}
\arguments{
\item{x}{A vector of lines from the Latex file}
}
\value{
A vector of lines with latex tables replace by something
that will render a Rmd table
}
\description{
Process the tables in a Latex file vector and
put into Rmd format
}
