% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTabulars.r
\name{processTabulars}
\alias{processTabulars}
\title{processTables}
\usage{
processTabulars(x, tabInbed = TRUE)
}
\arguments{
\item{x}{A vector of lines from the Latex file}

\item{tabInbed}{logical if a tabular within other environments, 
with default value "tabInbed = TRUE"}
}
\value{
A vector of lines with latex tables replace by something
that will render a Rmd table
}
\description{
Process the tables in a Latex file vector and
put into Rmd format
}
