% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBibliography.r
\name{readBibliography}
\alias{readBibliography}
\title{readBibliography}
\usage{
readBibliography(bblFile, style)
}
\arguments{
\item{bblFile}{The bbl file to read and format}

\item{style}{A text string specifying the style of the bibliography.
Only difference this makes is that for style == "plain", the '[n]'
numbers are added at the beginning.}
}
\value{
A vector of character strings containing the
formatted bibliography
}
\description{
Reads the .bbl file from a tex document and
formats the bibliography for insertion into a .Rmd document
}
