% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBib.r
\name{parseBib}
\alias{parseBib}
\title{parseBib}
\usage{
parseBib(x)
}
\arguments{
\item{x}{A latex bibitem entry from a Latex .bbl file}
}
\value{
A string containing the bibitem
}
\description{
Parse a bibitem and return the bibliography item
}
\author{
Trent McDonald
}
