% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTex.r
\name{parseTex}
\alias{parseTex}
\title{parseTex}
\usage{
parseTex(x, argNum, delims = "{}")
}
\arguments{
\item{x}{A latex command, like "\\bibcite{text}{text}"}

\item{argNum}{The number of the argument desired.}

\item{delims}{The delimiters to use.  This must be a single
string, and only the first two characters are used.  First character
is the opening delimiter, second character is closing delimiter.}
}
\value{
A string, the desired argument. If the argument is not found,
the empty string is returned and a warning is fired.
}
\description{
Parse a Tex command and return a specified argument.
}
\examples{
tex <- "command{1}{1984}{{Albert and Anderson}}{{}}"
parseTex(tex, 1)
parseTex(tex, 3)

}
\author{
Trent McDonald
}
