% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBibliography.r
\name{readBibliography}
\alias{readBibliography}
\title{readBibliography}
\usage{
readBibliography(bblFile)
}
\arguments{
\item{bblFile}{The bbl file to read and format}
}
\value{
A vector of character strings containing the
formatted bibliography
}
\description{
Reads the .bbl file from a tex document and
formats the bibliography for insertion into a .Rmd document
}
