% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCitations.r
\name{readCitations}
\alias{readCitations}
\title{readCitations}
\usage{
readCitations(auxFile)
}
\description{
Reads the .AUX file and create a citation table so that
citations can be substituted into the translated .Rmd file.
}
