% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex2rmd.r
\name{tex2rmd}
\alias{tex2rmd}
\title{tex2rmd}
\usage{
tex2rmd(infile)
}
\arguments{
\item{infile}{Full path to the input Latex file.}
}
\value{
The markdown code is written to a file named <root>.Rmd,
where \code{inFile} is <root>.tex.  The markdown code in the
form of a vector, one line per element, is invisably returned.
}
\description{
Converts a raw LaTex file to RMarkdown format, then to Word
format.  Can optionally convert to any format supported by RMarkdown
}
\details{
The general workflow to convert a Latex document into Markdown,
and eventually Word, is as follows:
\enumerate{
  \item Compile the latex document, using pdftex or whatever,
  so that the .aux and .bbl files are generated.
  \item Run \code{tex3rmd} specifying the file containing
  the raw Latex code as the input.  The associated .aux
  and .bbl files must be in the same directory. This will
  generate a markdown document of the same root name but
  with .Rmd extension.
  \item Compile the .Rmd file.  In RStudio, open it and hit <ctrl>-<shift>-K.
}
}
\examples{
\\notrun{
tex2rmd("main.tex")
 }
}
\author{
Trent McDonald
}
